popup/views/RecoverAccount: reskin RecoverAccount (#360)
* reskin RecoverAccount

* move to scss, fix textarea input

* cleanup
